Implémenter dans le Backend l'Endpoint pour récupérer la liste des emprunteurs

+++ backend/src/main/java/com/example/GestionBibliotheque/Repositories/LivreRepository.java
@@ -8,6 +8,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface LivreRepository extends JpaRepository<Livre, Long> {
+public interface LivreRepository extends JpaRepository<Livre, Integer> {
     List<Livre> findAllByOrderByAuteurAsc();
 }
